Move `graphpublisher` classes to `graph-publisher`
