+++ library/plugins/graph/graph-publisher/src/main/java/com/cdsap/talaiot/publisher/graph/GexfPublisher.kt
@@ -1,9 +1,9 @@
-package com.cdsap.talaiot.publisher.graphpublisher
+package com.cdsap.talaiot.publisher.graph
 
 import com.cdsap.talaiot.entities.ExecutionReport
 import com.cdsap.talaiot.entities.TaskMessageState
 import com.cdsap.talaiot.logger.LogTracker
-import com.cdsap.talaiot.publisher.graphpublisher.resources.ResourcesGexf
+import com.cdsap.talaiot.publisher.graph.resources.ResourcesGexf
 import com.cdsap.talaiot.writer.FileWriter
 import java.util.concurrent.Executor
