+++ library/plugins/graph/graph-publisher/src/main/java/com/cdsap/talaiot/publisher/graph/GraphPublisherType.kt
@@ -1,4 +1,4 @@
-package com.cdsap.talaiot.publisher.graphpublisher
+package com.cdsap.talaiot.publisher.graph
 
 /**
  * Represents the default Publishers for Graphs
