+++ library/plugins/graph/graph-publisher/src/main/java/com/cdsap/talaiot/publisher/graph/HtmlPublisher.kt
@@ -1,10 +1,10 @@
-package com.cdsap.talaiot.publisher.graphpublisher
+package com.cdsap.talaiot.publisher.graph
 
 import com.cdsap.talaiot.entities.ExecutionReport
 import com.cdsap.talaiot.entities.TaskMessageState
 import com.cdsap.talaiot.logger.LogTracker
-import com.cdsap.talaiot.publisher.graphpublisher.resources.ResourcesHtml
-import com.cdsap.talaiot.publisher.graphpublisher.resources.ResourcesHtml.LEGEND_HEADER
+import com.cdsap.talaiot.publisher.graph.resources.ResourcesHtml
+import com.cdsap.talaiot.publisher.graph.resources.ResourcesHtml.LEGEND_HEADER
 import com.cdsap.talaiot.writer.FileWriter
 import java.util.concurrent.Executor
