+++ library/plugins/graph/graph-publisher/src/main/java/com/cdsap/talaiot/publisher/graph/resources/ResourcesGexf.kt
@@ -1,4 +1,4 @@
-package com.cdsap.talaiot.publisher.graphpublisher.resources
+package com.cdsap.talaiot.publisher.graph.resources
 
 /**
  * HEADER and VALUE constants used by the GexfPublisher
