+++ library/plugins/talaiot-legacy/build.gradle.kts
@@ -15,7 +15,7 @@ dependencies {
     implementation(project(":library:talaiot"))
     implementation(project(":library:plugins:base:base-publisher"))
     implementation(project(":library:plugins:hybrid:hybrid-publisher"))
-    implementation("guru.nidi:graphviz-java:0.8.3")
+    implementation(project(":library:plugins:graph:graph-publisher"))
     implementation("org.influxdb:influxdb-java:2.19")
     implementation("com.github.oshi:oshi-core:3.13.3")
     implementation("com.google.code.gson:gson:2.8.5")
