+++ library/plugins/talaiot-legacy/src/main/kotlin/com/cdsap/talaiot/plugin/TalaiotConfigurationProvider.kt
@@ -6,8 +6,9 @@ import com.cdsap.talaiot.provider.PublisherConfigurationProvider
 import com.cdsap.talaiot.publisher.*
 import com.cdsap.talaiot.publisher.JsonPublisher
 import com.cdsap.talaiot.publisher.OutputPublisher
-import com.cdsap.talaiot.publisher.graphpublisher.GraphPublisherFactoryImpl
 import com.cdsap.talaiot.publisher.hybrid.HybridPublisher
+import com.cdsap.talaiot.publisher.graph.GraphPublisherFactoryImpl
+import com.cdsap.talaiot.publisher.graph.TaskDependencyGraphPublisher
 import com.cdsap.talaiot.publisher.pushgateway.PushGatewayFormatter
 import com.cdsap.talaiot.publisher.pushgateway.PushGatewayPublisher
 import com.cdsap.talaiot.publisher.rethinkdb.RethinkDbPublisher
