+++ library/talaiot/build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 dependencies {
     api(project(":library:talaiot-logger"))
     api(project(":library:talaiot-request"))
-    implementation("guru.nidi:graphviz-java:0.8.3")
+    api("guru.nidi:graphviz-java:0.8.3")
     implementation("org.influxdb:influxdb-java:2.19")
     implementation("com.github.oshi:oshi-core:3.13.3")
     implementation("com.google.code.gson:gson:2.8.5")
